Update home UI

+++ NaTour/Natour2022/app/src/main/java/com/cinamidea/natour_2022/routes_util/Route.java
@@ -14,7 +14,7 @@ public class Route {
 
     private String level;
 
-    private float duration;
+    private int duration;
 
     private int report_count;
 
@@ -34,7 +34,7 @@ public class Route {
 
     }
 
-    public Route(String name, String description, String creator_username, String level, float duration, int report_count, boolean disability_access, List<LatLng> coordinates, String tags, String image_base64, float length) {
+    public Route(String name, String description, String creator_username, String level, int duration, int report_count, boolean disability_access, List<LatLng> coordinates, String tags, String image_base64, float length) {
         this.name = name;
         this.description = description;
         this.creator_username = creator_username;
@@ -49,7 +49,7 @@ public class Route {
 
     }
 
-    public Route(String name, String description, String creator_username, String level, float duration, boolean disability_access, String image_base64,float length) {
+    public Route(String name, String description, String creator_username, String level, int duration, boolean disability_access, String image_base64,float length) {
         this.name = name;
         this.description = description;
         this.creator_username = creator_username;
@@ -98,11 +98,11 @@ public class Route {
         this.level = level;
     }
 
-    public float getDuration() {
+    public int getDuration() {
         return duration;
     }
 
-    public void setDuration(float duration) {
+    public void setDuration(int duration) {
         this.duration = duration;
     }
